DRY conversion RoaringBitmap's IntIterator to fastutil's IntIterator
