+++ processing/src/main/java/io/druid/segment/incremental/IncrementalIndexAdapter.java
@@ -36,8 +36,8 @@ import io.druid.segment.data.EmptyIndexedInts;
 import io.druid.segment.data.Indexed;
 import io.druid.segment.data.IndexedInts;
 import io.druid.segment.data.ListIndexed;
+import it.unimi.dsi.fastutil.ints.IntIterator;
 import org.joda.time.Interval;
-import org.roaringbitmap.IntIterator;
 
 import java.io.IOException;
 import java.util.Iterator;
@@ -286,44 +286,9 @@ public class IncrementalIndexAdapter implements IndexableAdapter
     }
 
     @Override
-    public it.unimi.dsi.fastutil.ints.IntIterator iterator()
+    public IntIterator iterator()
     {
-      return new Iter();
-    }
-
-    private class Iter implements it.unimi.dsi.fastutil.ints.IntIterator
-    {
-      final IntIterator baseIter = bitmapIndex.iterator();
-
-      @Override
-      public boolean hasNext()
-      {
-        return baseIter.hasNext();
-      }
-
-      @Override
-      public Integer next()
-      {
-        return baseIter.next();
-      }
-
-      @Override
-      public int nextInt()
-      {
-        return baseIter.next();
-      }
-
-      @Override
-      public void remove()
-      {
-        throw new UnsupportedOperationException();
-      }
-
-      @Override
-      public int skip(int n)
-      {
-        return IntIteratorUtils.skip(this, n);
-      }
+      return IntIteratorUtils.fromRoaringBitmapIntIterator(bitmapIndex.iterator());
     }
 
     @Override
